Extract scope handles from tree module
