+++ tools/enforcement-fixtures/overlays/layering-scope-driver-below/crates/shelterwood/src/scope.rs
@@ -1,0 +1,3 @@
+// Known-bad fixture: the shared scope-handle layer must not reference the
+// driver layer.
+use crate::driver::SystemRun;
