+++ tools/enforcement-fixtures/skeleton/crates/shelterwood/src/scope.rs
@@ -1,0 +1,4 @@
+// Known-good fixture for the scope-handle layer: near misses must not match
+// the word-bounded upward-layering pattern.
+const DRIVERISH: &str = "driver";
+const TREEHOUSE: &str = "tree";
